requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary>=2.9.5
pandas>=1.5.0
sqlalchemy>=1.4.0
openpyxl>=3.0.0
python-dotenv>=1.0.0
fastapi
uvicorn
passlib[bcrypt]
python-jose[cryptography]
python-multipart
google-generativeai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
keyring
SQLAlchemy 